Update Beauty of Nature.html
<!DOCTYPE html>
<html>
<head>
   <title> PHOTOGRAPHY WITH  NATURE</title>
</head>
 
<body>
    <h1> Beauty of Nature </h1>
    <p> Excellence of Nature

At the point when we notice and feel the excellence of nature it has such a quiet and mitigating impact on our whole self. It lastingly affects our faculties and gives an astonishing sensation of reward and enjoyment. At the point when we consider the magnificence of nature we promptly consider the grand and visual excellence of nature, similar to the taking off mountains, serene waterways, plants of trees, shades of rainbow, butterflies and birds, and the blooming blossoms. All that about nature provides us with the sensation of joy.

Nature is wonderful not on the grounds that it addresses our few issues but since of the way this multitude of normal structures render the sensation of euphoria and assist us with interfacing with our actual selves. The experience of this regular magnificence assorted and serious is the core of nature. All aspects of nature is wonderful and extraordinary in itself. Each and every second God is painting the new stroke of excellence in nature. Nature's excellence is uncovered wherever around us constantly. Morning dew on the velvet green grass, the blossoms tumbling from the tree in the lawn, the sparrow fabricating its home on the tree, butterflies to a great extent - everything is nature's excellence. Nature is alive, expressive and productive .</p>
</body>
 
</html>